+++ db/migrate/20190218145124_remove_event_attendance_table.rb
@@ -1,0 +1,7 @@
+class RemoveEventAttendanceTable < ActiveRecord::Migration[5.2]
+  def change
+    drop_table :event_attendances
+
+    add_reference :attendances, :event, foreign_key: true
+  end
+end