+++ db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 2019_02_15_231315) do
+ActiveRecord::Schema.define(version: 2019_02_18_145124) do
 
   # These are extensions that must be enabled in order to support this database
   enable_extension "plpgsql"
@@ -25,15 +25,8 @@ ActiveRecord::Schema.define(version: 2019_02_15_231315) do
     t.string "check_in"
     t.datetime "created_at", null: false
     t.datetime "updated_at", null: false
-  end
-
-  create_table "event_attendances", force: :cascade do |t|
     t.bigint "event_id"
-    t.bigint "attendance_id"
-    t.datetime "created_at", null: false
-    t.datetime "updated_at", null: false
-    t.index ["attendance_id"], name: "index_event_attendances_on_attendance_id"
-    t.index ["event_id"], name: "index_event_attendances_on_event_id"
+    t.index ["event_id"], name: "index_attendances_on_event_id"
   end
 
   create_table "events", force: :cascade do |t|
@@ -67,8 +60,7 @@ ActiveRecord::Schema.define(version: 2019_02_15_231315) do
     t.index ["username"], name: "index_users_on_username", unique: true
   end
 
-  add_foreign_key "event_attendances", "attendances"
-  add_foreign_key "event_attendances", "events"
+  add_foreign_key "attendances", "events"
   add_foreign_key "user_events", "events"
   add_foreign_key "user_events", "users"
 end