+++ spec/factories/user.rb
@@ -1,0 +1,8 @@
+# frozen_string_literal: true
+
+FactoryBot.define do
+  factory :user do
+    sequence(:email) { |n| "email#{n}@email.com" }
+    sequence(:username) { |n| "user#{n}" }
+  end
+end